requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
itsdangerous==2.1.2
python-dotenv==1.0.0
requests==2.32.3
Werkzeug==2.3.7
